test: dedupe pollUntil into smoke-test-utils; hermetic logDir in ruby integration smoke
Review follow-ups on #317: the new e2e test added a third verbatim copy
of pollUntil — promote it to smoke-test-utils.ts and point all three
suites at it. The ruby integration smoke test's logDir was
<cwd>/logs/tests, which buildAdapterCommand now really writes into
(sync prelude); use a per-test temp dir so the test stays hermetic and
the -r assertion never depends on repo-tree writability.
