+++ tests/e2e/mcp-server-break-on-exceptions.test.ts
@@ -26,7 +26,7 @@ import { spawn, execSync, type ChildProcess } from 'child_process';
 import net from 'net';
 import { Client } from '@modelcontextprotocol/sdk/client/index.js';
 import { StdioClientTransport } from '@modelcontextprotocol/sdk/client/stdio.js';
-import { parseSdkToolResult, callToolSafely } from './smoke-test-utils.js';
+import { parseSdkToolResult, callToolSafely, pollUntil } from './smoke-test-utils.js';
 import { prepareJavaExample } from './java-example-utils.js';
 import { prepareRustExample } from './rust-example-utils.js';
 import { skipIfSpawnBlocked } from '../test-utils/helpers/adapter-spawn.js';
@@ -61,20 +61,6 @@ async function getSessionSnapshot(client: Client, sessionId: string): Promise<Se
   const res = parseSdkToolResult(await client.callTool({ name: 'list_debug_sessions', arguments: {} }));
   const sessions = (res.sessions ?? []) as Array<SessionSnapshot & { id: string }>;
   return sessions.find(s => s.id === sessionId);
-}
-
-async function pollUntil<T>(
-  fn: () => Promise<T | undefined>,
-  timeoutMs: number,
-  intervalMs = 250
-): Promise<T | undefined> {
-  const deadline = Date.now() + timeoutMs;
-  for (;;) {
-    const value = await fn();
-    if (value !== undefined) return value;
-    if (Date.now() > deadline) return undefined;
-    await new Promise(r => setTimeout(r, intervalMs));
-  }
 }
 
 async function findFreePort(): Promise<number> {
